+++ src/tx.rs
@@ -14,20 +14,13 @@ use std::sync::Arc;
 use std::sync::RwLock;
 use std::time;
 
-use crate::btree_ascend;
-use crate::btree_ascend_greater_or_equal;
-use crate::btree_ascend_less_than;
-use crate::btree_ascend_range;
-use crate::btree_descend;
-use crate::btree_descend_greater_than;
-use crate::btree_descend_less_or_equal;
-use crate::btree_descend_range;
+use crate::btree_helpers::*;
 use crate::exps_compare_fn;
+use crate::item::DbItem;
+use crate::item::DbItemOpts;
 use crate::keys_compare_fn;
 use crate::Db;
 use crate::DbError;
-use crate::DbItem;
-use crate::DbItemOpts;
 use crate::Index;
 use crate::IndexOptions;
 use crate::LessFn;
